code/backend/round1Selector/requirements.txt
python-dotenv
semantic-kernel